refactor the remaining lobby commands
