+++ app/rooms/custom-lobby-room.ts
@@ -8,12 +8,12 @@ import {
 import { Dispatcher } from "@colyseus/command"
 import LobbyState from "./states/lobby-state"
 import { connect } from "mongoose"
+import BannedUser from "../models/mongo-models/banned-user"
 import ChatV2 from "../models/mongo-models/chat-v2"
 import UserMetadata from "../models/mongo-models/user-metadata"
-import BannedUser from "../models/mongo-models/banned-user"
 import { ILeaderboardInfo } from "../models/colyseus-models/leaderboard-info"
 import admin from "firebase-admin"
-import { MessageEmbed, WebhookClient } from "discord.js"
+import { WebhookClient } from "discord.js"
 import BotV2, { IBot } from "../models/mongo-models/bot-v2"
 import Meta, { IMeta } from "../models/mongo-models/meta"
 import ItemsStatistic, {
@@ -23,13 +23,7 @@ import PokemonsStatistic, {
   IPokemonsStatistic
 } from "../models/mongo-models/pokemons-statistic"
 import { PastebinAPI } from "pastebin-ts/dist/api"
-import { getAvatarSrc } from "../public/src/utils"
-import {
-  Emotion,
-  Transfer,
-  Title,
-  Role
-} from "../types"
+import { Emotion, Transfer, Title, Role } from "../types"
 import { nanoid } from "nanoid"
 import { logger } from "../utils/logger"
 import {
@@ -48,7 +42,13 @@ import {
   BuyBoosterCommand,
   OnSearchCommand,
   OnSearchByIdCommand,
-  ChangeAvatarCommand
+  ChangeAvatarCommand,
+  BanUserCommand,
+  UnbanUserCommand,
+  AddBotCommand,
+  DeleteBotCommand,
+  OnBotUploadCommand,
+  createBotList
 } from "./commands/lobby-commands"
 
 export default class CustomLobbyRoom extends Room<LobbyState> {
@@ -169,198 +169,26 @@ export default class CustomLobbyRoom extends Room<LobbyState> {
     })
 
     this.onMessage(Transfer.DELETE_BOT_DATABASE, async (client, message) => {
-      try {
-        const user = this.state.users.get(client.auth.uid)
-        if (
-          user &&
-          (user.role === Role.ADMIN ||
-            user.role === Role.BOT_MANAGER ||
-            user.role === Role.MODERATOR)
-        ) {
-          const id = message
-          const botData = this.bots.get(id)
-          client.send(
-            Transfer.BOT_DATABASE_LOG,
-            `deleting bot ${botData?.name}by @${botData?.author} id ${id}`
-          )
-          const resultDelete = await BotV2.deleteOne({ id: id })
-          client.send(
-            Transfer.BOT_DATABASE_LOG,
-            JSON.stringify(resultDelete, null, 2)
-          )
-          const dsEmbed = new MessageEmbed()
-            .setTitle(
-              `BOT ${botData?.name} by @${botData?.author} deleted by ${user.name}`
-            )
-            .setAuthor({
-              name: user.name,
-              iconURL: getAvatarSrc(user.avatar)
-            })
-            .setDescription(
-              `BOT ${botData?.name} by @${botData?.author} (id: ${message} ) deleted by ${user.name}`
-            )
-            .setThumbnail(getAvatarSrc(botData?.avatar ? botData?.avatar : ""))
-          try {
-            this.discordWebhook?.send({
-              embeds: [dsEmbed]
-            })
-          } catch (error) {
-            logger.error(error)
-          }
-
-          this.bots.delete(id)
-          this.broadcast(Transfer.REQUEST_BOT_LIST, this.createBotList())
-        }
-      } catch (error) {
-        logger.error(error)
-        client.send(Transfer.BOT_DATABASE_LOG, JSON.stringify(error))
-      }
+      this.dispatcher.dispatch(new DeleteBotCommand(), { client, message })
     })
 
     this.onMessage(Transfer.ADD_BOT_DATABASE, async (client, message) => {
-      try {
-        const user = this.state.users.get(client.auth.uid)
-        if (
-          user &&
-          (user.role === Role.ADMIN ||
-            user.role === Role.BOT_MANAGER ||
-            user.role === Role.MODERATOR)
-        ) {
-          const id = message.slice(21)
-          client.send(Transfer.BOT_DATABASE_LOG, `retrieving id : ${id} ...`)
-          client.send(Transfer.BOT_DATABASE_LOG, "retrieving data ...")
-          const data = await this.pastebin?.getPaste(id, false)
-          if (data) {
-            client.send(Transfer.BOT_DATABASE_LOG, "parsing JSON data ...")
-            const json = JSON.parse(data)
-            const resultDelete = await BotV2.deleteMany({
-              avatar: json.avatar,
-              author: json.author
-            })
-            const keys = new Array<string>()
-            this.bots.forEach((b) => {
-              if (b.avatar === json.avatar && b.author === json.author) {
-                keys.push(b.id)
-              }
-            })
-            keys.forEach((k) => {
-              this.bots.delete(k)
-            })
-            client.send(
-              Transfer.BOT_DATABASE_LOG,
-              JSON.stringify(resultDelete, null, 2)
-            )
-            client.send(
-              Transfer.BOT_DATABASE_LOG,
-              `creating Bot ${json.avatar} by ${json.author}...`
-            )
-            const resultCreate = await BotV2.create({
-              name: json.name,
-              avatar: json.avatar,
-              elo: json.elo ? json.elo : 1200,
-              author: json.author,
-              steps: json.steps,
-              id: nanoid()
-            })
-
-            const dsEmbed = new MessageEmbed()
-              .setTitle(
-                `BOT ${json.name} by @${json.author} loaded by ${user.name}`
-              )
-              .setURL(message as string)
-              .setAuthor({
-                name: user.name,
-                iconURL: getAvatarSrc(user.avatar)
-              })
-              .setDescription(
-                `BOT ${json.name} by @${json.author} (url: ${message} ) loaded by ${user.name}`
-              )
-              .setThumbnail(getAvatarSrc(json.avatar))
-            try {
-              this.discordWebhook?.send({
-                embeds: [dsEmbed]
-              })
-            } catch (error) {
-              logger.error(error)
-            }
-
-            this.bots.set(resultCreate.id, resultCreate)
-            this.broadcast(Transfer.REQUEST_BOT_LIST, this.createBotList())
-          } else {
-            client.send(
-              Transfer.BOT_DATABASE_LOG,
-              `no pastebin found with given url ${message}`
-            )
-          }
-        }
-      } catch (error) {
-        logger.error(error)
-        client.send(Transfer.BOT_DATABASE_LOG, JSON.stringify(error))
-      }
+      this.dispatcher.dispatch(new AddBotCommand(), { client, message })
     })
 
-    this.onMessage(Transfer.UNBAN, (client, message) => {
-      try {
-        const user = this.state.users.get(client.auth.uid)
-        if (
-          user &&
-          (user.role === Role.ADMIN || user.role === Role.MODERATOR)
-        ) {
-          BannedUser.deleteOne({ uid: message.uid }, (err, res) => {
-            if (err) {
-              logger.error
-            }
-            if (res?.deletedCount > 0) {
-              client.send(
-                Transfer.BANNED,
-                `${user.name} unbanned the user ${message.name}`
-              )
-            }
-          })
-        }
-      } catch (error) {
-        logger.error(error)
+    this.onMessage(
+      Transfer.UNBAN,
+      (client, { uid, name }: { uid: string; name: string }) => {
+        this.dispatcher.dispatch(new UnbanUserCommand(), { client, uid, name })
       }
-    })
+    )
 
-    this.onMessage(Transfer.BAN, (client, message) => {
-      try {
-        const user = this.state.users.get(client.auth.uid)
-        if (
-          user &&
-          (user.role === Role.ADMIN || user.role === Role.MODERATOR)
-        ) {
-          this.state.removeMessages(message.uid)
-          BannedUser.findOne({ uid: message.uid }, (err, banned) => {
-            if (err) {
-              logger.error(err)
-            }
-            if (!banned) {
-              BannedUser.create({
-                uid: message.uid,
-                author: user.name,
-                time: Date.now(),
-                name: message.name
-              })
-              client.send(
-                Transfer.BANNED,
-                `${user.name} banned the user ${message.name}`
-              )
-            } else {
-              client.send(Transfer.BANNED, `${message.name} was already banned`)
-            }
-          })
-          this.clients.forEach((c) => {
-            if (c.auth.uid === message.uid) {
-              c.send(Transfer.BAN)
-              c.leave()
-            }
-          })
-        }
-      } catch (error) {
-        logger.error(error)
+    this.onMessage(
+      Transfer.BAN,
+      (client, { uid, name }: { uid: string; name: string }) => {
+        this.dispatcher.dispatch(new BanUserCommand(), { client, uid, name })
       }
-    })
+    )
 
     this.onMessage(Transfer.NEW_MESSAGE, (client, message) => {
       this.dispatcher.dispatch(new OnNewMessageCommand(), { client, message })
@@ -404,49 +232,13 @@ export default class CustomLobbyRoom extends Room<LobbyState> {
       }
     )
 
-    this.onMessage(Transfer.BOT_CREATION, (client, message) => {
-      try {
-        const bot = message.bot
-        const user = this.state.users.get(client.auth.uid)
-        if (!user) return
-        this.pastebin
-          ?.createPaste({
-            text: JSON.stringify(bot),
-            title: `${user.name} has uploaded BOT ${bot.name}`,
-            format: "json"
-          })
-          .then((data: unknown) => {
-            const dsEmbed = new MessageEmbed()
-              .setTitle(`BOT ${bot.name} created by ${bot.author}`)
-              .setURL(data as string)
-              .setAuthor({
-                name: user.name,
-                iconURL: getAvatarSrc(user.avatar)
-              })
-              .setDescription(
-                `A new bot has been created by ${user.name}, You can import the data in the Pokemon Auto Chess Bot Builder (url: ${data} ).`
-              )
-              .setThumbnail(getAvatarSrc(bot.avatar))
-            client.send(Transfer.PASTEBIN_URL, { url: data as string })
-            try {
-              this.discordWebhook?.send({
-                embeds: [dsEmbed]
-              })
-            } catch (error) {
-              logger.error(error)
-            }
-          })
-          .catch((error) => {
-            logger.error(error)
-          })
-      } catch (error) {
-        logger.error(error)
-      }
+    this.onMessage(Transfer.BOT_CREATION, (client, { bot }: { bot: IBot }) => {
+      this.dispatcher.dispatch(new OnBotUploadCommand(), { client, bot })
     })
 
     this.onMessage(Transfer.REQUEST_BOT_LIST, (client) => {
       try {
-        client.send(Transfer.REQUEST_BOT_LIST, this.createBotList())
+        client.send(Transfer.REQUEST_BOT_LIST, createBotList(this.bots))
       } catch (error) {
         logger.error(error)
       }
@@ -546,9 +338,18 @@ export default class CustomLobbyRoom extends Room<LobbyState> {
       Transfer.CHANGE_AVATAR,
       (
         client,
-        { index, emotion, shiny }: { index: string; emotion: Emotion; shiny: boolean }
+        {
+          index,
+          emotion,
+          shiny
+        }: { index: string; emotion: Emotion; shiny: boolean }
       ) => {
-        this.dispatcher.dispatch(new ChangeAvatarCommand(), { client, index, emotion, shiny })        
+        this.dispatcher.dispatch(new ChangeAvatarCommand(), {
+          client,
+          index,
+          emotion,
+          shiny
+        })
       }
     )
 
@@ -560,123 +361,9 @@ export default class CustomLobbyRoom extends Room<LobbyState> {
           if (err != null) {
             logger.error("Error connecting to Mongo", err)
           }
-          ChatV2.find(
-            { time: { $gt: Date.now() - 86400000 } },
-            (err, messages) => {
-              if (err) {
-                logger.error(err)
-              } else {
-                messages.forEach((message) => {
-                  this.state.addMessage(
-                    nanoid(),
-                    message.payload,
-                    message.authorId,
-                    message.author,
-                    message.avatar,
-                    message.time,
-                    false
-                  )
-                })
-              }
-            }
-          )
-          UserMetadata.find(
-            {},
-            ["displayName", "avatar", "elo"],
-            { limit: 100, sort: { elo: -1 } },
-            (err, users) => {
-              if (err) {
-                logger.error(err)
-              } else {
-                for (let i = 0; i < users.length; i++) {
-                  const user = users[i]
-                  this.leaderboard.push({
-                    name: user.displayName,
-                    rank: i + 1,
-                    avatar: user.avatar,
-                    value: user.elo
-                  })
-                }
-              }
-            }
-          )
-          UserMetadata.find(
-            {},
-            ["displayName", "avatar", "level"],
-            { limit: 100, sort: { level: -1 } },
-            (err, users) => {
-              if (err) {
-                logger.error(err)
-              } else {
-                for (let i = 0; i < users.length; i++) {
-                  const user = users[i]
-                  this.levelLeaderboard.push({
-                    name: user.displayName,
-                    rank: i + 1,
-                    avatar: user.avatar,
-                    value: user.level
-                  })
-                }
-              }
-            }
-          )
-          BotV2.find({}, {}, { sort: { elo: -1 } }, (_err, bots) => {
-            const ids = new Array<string>()
-            bots.forEach((bot, i) => {
-              if (ids.includes(bot.id)) {
-                const id = nanoid()
-                bot.id = id
-                bot.save()
-              }
-              ids.push(bot.id)
-              this.bots.set(bot.id, bot)
-              this.botLeaderboard.push({
-                name: `${bot.name} by @${bot.author}`,
-                avatar: bot.avatar,
-                rank: i + 1,
-                value: bot.elo
-              })
-            })
-          })
-          Meta.find(
-            {},
-            [
-              "cluster_id",
-              "count",
-              "ratio",
-              "winrate",
-              "mean_rank",
-              "types",
-              "pokemons"
-            ],
-            (err, docs) => {
-              if (err) {
-                logger.error(err)
-              } else {
-                docs.forEach((doc) => {
-                  this.meta.push(doc)
-                })
-              }
-            }
-          )
-          ItemsStatistic.find({}, (err, docs) => {
-            if (err) {
-              logger.error(err)
-            } else {
-              docs.forEach((doc) => {
-                this.metaItems.push(doc)
-              })
-            }
-          })
-          PokemonsStatistic.find({}, (err, docs) => {
-            if (err) {
-              logger.error(err)
-            } else {
-              docs.forEach((doc) => {
-                this.metaPokemons.push(doc)
-              })
-            }
-          })
+          this.fetchChat()
+          this.fetchLeaderboards()
+          this.fetchMeta()
         }
       )
       resolve()
@@ -727,22 +414,129 @@ export default class CustomLobbyRoom extends Room<LobbyState> {
     }
   }
 
-  createBotList() {
-    const botList = new Array<{
-      name: string
-      avatar: string
-      author: string
-      id: string
-    }>()
+  fetchChat(){
+    ChatV2.find(
+      { time: { $gt: Date.now() - 86400000 } },
+      (err, messages) => {
+        if (err) {
+          logger.error(err)
+        } else {
+          messages.forEach((message) => {
+            this.state.addMessage(
+              nanoid(),
+              message.payload,
+              message.authorId,
+              message.author,
+              message.avatar,
+              message.time,
+              false
+            )
+          })
+        }
+      }
+    )
+  }
 
-    this.bots.forEach((b) => {
-      botList.push({
-        name: b.name,
-        avatar: b.avatar,
-        id: b.id,
-        author: b.author
+  fetchMeta(){
+    Meta.find(
+      {},
+      [
+        "cluster_id",
+        "count",
+        "ratio",
+        "winrate",
+        "mean_rank",
+        "types",
+        "pokemons"
+      ],
+      (err, docs) => {
+        if (err) {
+          logger.error(err)
+        } else {
+          docs.forEach((doc) => {
+            this.meta.push(doc)
+          })
+        }
+      }
+    )
+    ItemsStatistic.find({}, (err, docs) => {
+      if (err) {
+        logger.error(err)
+      } else {
+        docs.forEach((doc) => {
+          this.metaItems.push(doc)
+        })
+      }
+    })
+    PokemonsStatistic.find({}, (err, docs) => {
+      if (err) {
+        logger.error(err)
+      } else {
+        docs.forEach((doc) => {
+          this.metaPokemons.push(doc)
+        })
+      }
+    })
+  }
+
+  fetchLeaderboards() {
+    UserMetadata.find(
+      {},
+      ["displayName", "avatar", "elo"],
+      { limit: 100, sort: { elo: -1 } },
+      (err, users) => {
+        if (err) {
+          logger.error(err)
+        } else {
+          for (let i = 0; i < users.length; i++) {
+            const user = users[i]
+            this.leaderboard.push({
+              name: user.displayName,
+              rank: i + 1,
+              avatar: user.avatar,
+              value: user.elo
+            })
+          }
+        }
+      }
+    )
+    UserMetadata.find(
+      {},
+      ["displayName", "avatar", "level"],
+      { limit: 100, sort: { level: -1 } },
+      (err, users) => {
+        if (err) {
+          logger.error(err)
+        } else {
+          for (let i = 0; i < users.length; i++) {
+            const user = users[i]
+            this.levelLeaderboard.push({
+              name: user.displayName,
+              rank: i + 1,
+              avatar: user.avatar,
+              value: user.level
+            })
+          }
+        }
+      }
+    )
+    BotV2.find({}, {}, { sort: { elo: -1 } }, (_err, bots) => {
+      const ids = new Array<string>()
+      bots.forEach((bot, i) => {
+        if (ids.includes(bot.id)) {
+          const id = nanoid()
+          bot.id = id
+          bot.save()
+        }
+        ids.push(bot.id)
+        this.bots.set(bot.id, bot)
+        this.botLeaderboard.push({
+          name: `${bot.name} by @${bot.author}`,
+          avatar: bot.avatar,
+          rank: i + 1,
+          value: bot.elo
+        })
       })
     })
-    return botList
   }
 }
